+++ app/src/contexts/alert/AlertTypes.ts
@@ -1,0 +1,15 @@
+export type AnimationState = 'closed' | 'opening' | 'open' | 'closing'
+
+export interface AlertContextValue {
+  showAlert: (message: string, buttonText: string, onClose?: () => void) => void
+  closeAlert: () => void
+  discardAlert: () => void
+}
+
+export interface AlertStateContextValue {
+  animationState: AnimationState
+  showContent: boolean
+  message: string
+  buttonText: string
+  isVisible: boolean
+}
