+++ app/src/contexts/drag-and-drop/DragAndDropContext.ts
@@ -1,0 +1,5 @@
+import { createContext } from 'react'
+import type { DragAndDropContextValue } from './DragAndDropTypes'
+
+export const DragAndDropContext = createContext<DragAndDropContextValue | undefined>(undefined)
+
