+++ app/src/contexts/drag-and-drop/DragAndDropTypes.ts
@@ -1,0 +1,3 @@
+export interface DragAndDropContextValue {
+  isDragging: boolean
+}
