+++ app/src/contexts/logger/LoggerContext.ts
@@ -1,0 +1,5 @@
+import { createContext } from 'react'
+import type { LoggerContextType } from './LoggerTypes'
+
+export const LoggerContext = createContext<LoggerContextType | null>(null)
+
