+++ app/src/contexts/rpc/RpcContext.ts
@@ -1,0 +1,4 @@
+import { createContext } from 'react'
+import type { AiutaAppRpc } from '@lib/rpc'
+
+export const RpcContext = createContext<AiutaAppRpc | undefined>(undefined)
