+++ app/src/contexts/rpc/RpcTypes.ts
@@ -1,0 +1,3 @@
+import type { AiutaAppRpc } from '@lib/rpc'
+
+export type { AiutaAppRpc }
