+++ app/src/contexts/rpc/useRpc.ts
@@ -1,26 +1,6 @@
-import React, { createContext, useContext, ReactNode } from 'react'
+import { useContext } from 'react'
 import type { AiutaAppRpc } from '@lib/rpc'
-import { useLoggerConfig } from '@/hooks'
-
-export const RpcContext = createContext<AiutaAppRpc | undefined>(undefined)
-
-interface RpcProviderProps {
-  children: ReactNode
-  rpc: AiutaAppRpc
-}
-
-export function RpcProvider({ children, rpc }: RpcProviderProps) {
-  return (
-    <RpcContext.Provider value={rpc}>
-      <RpcLoggerConfigWrapper>{children}</RpcLoggerConfigWrapper>
-    </RpcContext.Provider>
-  )
-}
-
-function RpcLoggerConfigWrapper({ children }: { children: React.ReactNode }) {
-  useLoggerConfig()
-  return <>{children}</>
-}
+import { RpcContext } from './RpcContext'
 
 /**
  * Creates a simple RPC stub for standalone mode
