+++ app/src/contexts/share/ShareContext.ts
@@ -1,0 +1,5 @@
+import { createContext } from 'react'
+import type { ShareContextType } from './ShareTypes'
+
+export const ShareContext = createContext<ShareContextType | null>(null)
+
