+++ app/src/contexts/share/ShareTypes.ts
@@ -1,0 +1,10 @@
+export type AnimationState = 'closed' | 'opening' | 'open' | 'closing'
+
+export interface ShareContextType {
+  modalData: { imageUrl: string } | null
+  animationState: AnimationState
+  isOpen: boolean
+  isVisible: boolean
+  openShareModal: (imageUrl: string) => void
+  closeShareModal: () => void
+}
