+++ app/src/contexts/storage/StorageContext.ts
@@ -1,0 +1,5 @@
+import { createContext } from 'react'
+import type { StorageContextValue } from './StorageTypes'
+
+export const StorageContext = createContext<StorageContextValue | null>(null)
+
