+++ app/src/contexts/storage/StorageTypes.ts
@@ -1,0 +1,6 @@
+import type { IStorageBackend } from '@/utils/storage/backends/IStorageBackend'
+
+export interface StorageContextValue {
+  backend: IStorageBackend
+}
+
